refactor: improved CartItemCard structure
